separating the lowlevel extension functions which would be useful for client customisations, but won't be needed by standard clients
