+++ .editorconfig
@@ -1,0 +1,2 @@
+[*.{kt,kts}]
+max_line_length=125
