+++ n8-core/src/main/kotlin/co/early/n8/RestrictedNavigation.kt
@@ -9,39 +9,39 @@ import co.early.n8.RestrictedNavigation.NotEndNode
  * during library development, and that's what these classes are for
  */
 internal class RestrictedNavigation {
-    internal sealed class NotBackStack<L: Any, T: Any> {
-        internal data class IsEndNode<L: Any, T: Any>(val value: Navigation.EndNode<L, T>) : NotBackStack<L, T>()
-        internal data class IsTabHost<L: Any, T: Any>(val value: Navigation.TabHost<L, T>) : NotBackStack<L, T>()
+    internal sealed class NotBackStack<L : Any, T : Any> {
+        internal data class IsEndNode<L : Any, T : Any>(val value: Navigation.EndNode<L, T>) : NotBackStack<L, T>()
+        internal data class IsTabHost<L : Any, T : Any>(val value: Navigation.TabHost<L, T>) : NotBackStack<L, T>()
     }
 
-    internal sealed class NotEndNode<L: Any, T: Any> {
-        internal data class IsBackStack<L: Any, T: Any>(val value: Navigation.BackStack<L, T>) : NotEndNode<L, T>()
-        internal data class IsTabHost<L: Any, T: Any>(val value: Navigation.TabHost<L, T>) : NotEndNode<L, T>()
+    internal sealed class NotEndNode<L : Any, T : Any> {
+        internal data class IsBackStack<L : Any, T : Any>(val value: Navigation.BackStack<L, T>) : NotEndNode<L, T>()
+        internal data class IsTabHost<L : Any, T : Any>(val value: Navigation.TabHost<L, T>) : NotEndNode<L, T>()
     }
 }
 
-internal fun <L: Any, T: Any> Navigation<L, T>.isBackStack(): Navigation.BackStack<L, T> {
+internal fun <L : Any, T : Any> Navigation<L, T>.isBackStack(): Navigation.BackStack<L, T> {
     return when (this) {
         is Navigation.BackStack<L, T> -> this
         is Navigation.EndNode, is Navigation.TabHost -> throw RuntimeException(errorMsg)
     }
 }
 
-internal fun <L: Any, T: Any> Navigation<L, T>.isEndNode(): Navigation.EndNode<L, T> {
+internal fun <L : Any, T : Any> Navigation<L, T>.isEndNode(): Navigation.EndNode<L, T> {
     return when (this) {
         is Navigation.EndNode -> this
         is Navigation.BackStack, is Navigation.TabHost -> throw RuntimeException(errorMsg)
     }
 }
 
-internal fun <L: Any, T: Any> Navigation<L, T>.isTabHost(): Navigation.TabHost<L, T> {
+internal fun <L : Any, T : Any> Navigation<L, T>.isTabHost(): Navigation.TabHost<L, T> {
     return when (this) {
         is Navigation.TabHost<L, T> -> this
         is Navigation.EndNode, is Navigation.BackStack -> throw RuntimeException(errorMsg)
     }
 }
 
-internal fun <L: Any, T: Any> Navigation<L, T>.notBackStack(): NotBackStack<L, T> {
+internal fun <L : Any, T : Any> Navigation<L, T>.notBackStack(): NotBackStack<L, T> {
     return when (this) {
         is Navigation.BackStack<L, T> -> throw RuntimeException(errorMsg)
         is Navigation.EndNode -> NotBackStack.IsEndNode(this)
@@ -49,7 +49,7 @@ internal fun <L: Any, T: Any> Navigation<L, T>.notBackStack(): NotBackStack<L, T
     }
 }
 
-internal fun <L: Any, T: Any> Navigation<L, T>.notEndNode(): NotEndNode<L, T> {
+internal fun <L : Any, T : Any> Navigation<L, T>.notEndNode(): NotEndNode<L, T> {
     return when (this) {
         is Navigation.BackStack -> NotEndNode.IsBackStack(this)
         is Navigation.EndNode -> throw RuntimeException(errorMsg)
