+++ n8-core/src/main/kotlin/co/early/n8/lowlevel/Ext.kt
@@ -1,0 +1,539 @@
+package co.early.n8.lowlevel
+
+import co.early.fore.kt.core.delegate.Fore
+import co.early.n8.Navigation
+import co.early.n8.Navigation.BackStack
+import co.early.n8.Navigation.EndNode
+import co.early.n8.Navigation.TabHost
+import co.early.n8.NavigationModel
+import co.early.n8.RestrictedNavigation
+import co.early.n8.RestrictedNavigation.NotBackStack
+import co.early.n8.TabBackMode
+import co.early.n8.TabHostLocation
+import co.early.n8.endNodeOf
+import co.early.n8.isBackStack
+import co.early.n8.notBackStack
+import co.early.n8.notEndNode
+
+/**
+ * The functions listed here are part of n8 library internals, some have been exposed publicly
+ * because they can be useful for clients who want to develop their own custom behaviours. This is
+ * advanced usage though which most clients aren't going to need.
+ *
+ * Unlike the functions is the other packages (which are designed to be impossible to break), the
+ * functions here are much more easy to misuse, please pay close attention to the warnings and
+ * explanations you'll find sprinkled about the source code.
+ *
+ * And remember: Fore.i(navigation.toString(diagnostics = true)) is your friend!
+ */
+
+// TODO some of these recursive functions work up the tree, some work down, is it worth highlighting this
+// or can we make it consistent?
+
+
+private const val padStep: Int = 4
+
+/**
+ * For a newly created navigation (especially one which has been copied and altered from an
+ * original) the children may not realise who their new parent is, this function corrects that
+ * information
+ */
+fun <L : Any, T : Any> Navigation<L, T>.populateChildParents(): Navigation<L, T> {
+    return when (this) {
+        is BackStack -> populateChildParents()
+        is EndNode -> this
+        is TabHost -> populateChildParents()
+    }
+}
+
+/**
+ * For a new TabHost (including one copied from an old TabHost with minor changes), each of the Tabs
+ * will contain a BackStack that does not realise who its new parent is, this function corrects that
+ * information
+ */
+internal fun <L : Any, T : Any> TabHost<L, T>.populateChildParents(): TabHost<L, T> {
+    for (backStack in tabs) {
+        backStack.parent = this
+        backStack.populateChildParents()
+    }
+    return this
+}
+
+/**
+ * For a new BackStack (including one copied from an old BackStack with minor changes), none of the
+ * items in the Stack will realise who their new parent is, this function corrects that
+ * information
+ */
+internal fun <L : Any, T : Any> BackStack<L, T>.populateChildParents(): BackStack<L, T> {
+    for (navigation in stack) {
+        when (val notBackStack = navigation.notBackStack()) {
+            is NotBackStack.IsEndNode -> notBackStack.value.parent = this
+            is NotBackStack.IsTabHost -> {
+                notBackStack.value.parent = this
+                notBackStack.value.populateChildParents()
+            }
+        }
+    }
+    return this
+}
+
+/**
+ * Mutates the navigation graph by replacing the oldItem with the newItem, applying the change
+ * all the way back up the chain, via parent references, finally returning the complete
+ * mutated navigation graph
+ *
+ * NOTE 1: populateParents() MUST be called on the navigation graph BEFORE calling this
+ * function initially (not required for subsequent recursive calls). This will already be the
+ * case provided the navigation graph has been constructed using the helper
+ * functions: backStackOf() tabHostOf() endNodeOf()
+ *
+ * NOTE 2: There are some restrictions placed on the mutation. Because of how a navigation
+ * graph is structured (the top level item is always a BackStack, a BackStack's direct children
+ * can only be EndNodes or TabHosts, a TabHost's direct children can only be BackStacks, it's
+ * often not valid to change the item _type_ as part of a mutation operation). When the
+ * mutated navigation item is re-integrated into the navigation graph in the same position,
+ * its new type can make it invalid. The only type change that is safe
+ * is to mutate a TabHost to an EndNode or vice versa. Any other type change will cause this
+ * function to immediately throw a runtime exception
+ *
+ * @oldItem any item in the navigation graph, often (but not necessarily) this will be the item
+ * at currentItem() as mutation travels backwards via parent references. The parent value must be
+ * valid as it will be adjusted to add the newItem as its child
+ * @newItem the replacement for the oldItem, this may have different values, children or even be of
+ * a different type (see above for restrictions). The parent value of the newItem will be ignored
+ * and a new value set as part of the mutation
+ * @ensureOnHistoryPath If ensureOnHistoryPath is true, the newItem will also be placed on the
+ * path to the currentItem if it isn't already. Relevant for cases when the oldItem is not on the
+ * history path but is still present somewhere in the navigation graph. Replacing oldItem with
+ * newItem will mutate the navigation graph, but it will not necessarily change the history path
+ * if the oldItem was not on the history path in the first place (which may or may not be what was
+ * intended). ensureOnHistoryPath = true ensures that once the mutation is complete, the newItem is
+ * accessible by the user on the back stack (or is the actual currentItem)
+ *
+ * @returns a new complete mutated navigation graph containing the newItem, with all parent and child
+ * references updated
+ */
+fun <L : Any, T : Any> mutateNavigation(
+    oldItem: Navigation<L, T>,
+    newItem: Navigation<L, T>,
+    ensureOnHistoryPath: Boolean = false,
+): Navigation<L, T> {
+
+    Fore.d("mutateNavigation() OLD:$oldItem NEW:$newItem ensureOnHistoryPath:$ensureOnHistoryPath")
+
+    val result = oldItem.parent?.let { oldItemParent -> // only the top level item has no parent
+        mutateNavigation(
+            oldItem = oldItemParent,
+            newItem = when (val oldParent =
+                oldItemParent.notEndNode()) { // EndNodes are NOT valid parents
+
+                is RestrictedNavigation.NotEndNode.IsBackStack -> {
+
+                    Fore.d("mutateNavigation()... oldParent is BackStack")
+
+                    oldParent.value.copy(
+                        stack = if (ensureOnHistoryPath) {
+                            oldParent.value.stack
+                                .takeWhile { it !== oldItem }
+                                .toMutableList()
+                                .also { it.add(newItem) }
+                        } else {
+                            oldParent.value.stack.map {
+                                if (it === oldItem) {
+                                    newItem
+                                } else it
+                            }
+                        }  // todo do we need a populatechildparents here?
+                    ).populateChildParents()
+                        .also { newParent ->  // all the entries in the parent stack need to reference their new parent
+                            newParent.stack.map {
+                                updateParent(
+                                    it,
+                                    newParent
+                                )
+                            }
+                        }
+                }
+
+                is RestrictedNavigation.NotEndNode.IsTabHost -> {
+
+                    Fore.d("mutateNavigation()... oldParent is Tabhost")
+
+                    val oldTabHostParent = oldParent.value
+
+                    var tabIndex = -1
+                    val newTabs = oldTabHostParent.tabs.mapIndexed { index, tab ->
+                        if (tab === oldItem) {
+                            tabIndex = index
+                            newItem.isBackStack()
+                        } else tab
+                    }
+
+                    require(tabIndex > -1) {
+                        "It should be impossible reach here, but if we do it's a bug [2]. Please file an issue, including the state of " +
+                                "the navigation graph just before the crash: " +
+                                "'navigationModel.toString(diagnostics = true)' and the operation performed"
+                    }
+
+                    val newHistory = if (ensureOnHistoryPath) {
+                        when (oldTabHostParent.tabBackModeDefault) {
+                            TabBackMode.Structural -> {
+                                listOf(tabIndex)
+                            }
+
+                            TabBackMode.Temporal -> {
+                                oldTabHostParent.selectedTabHistory.filter { tab ->
+                                    tab != tabIndex
+                                }.toMutableList().also { list ->
+                                    list.add(tabIndex)
+                                }
+                            }
+                        }
+                    } else {
+                        oldTabHostParent.selectedTabHistory
+                    }
+
+                    oldParent.value.copy(
+                        tabs = newTabs,
+                        selectedTabHistory = newHistory,
+                        // todo do we needa populatechildparents here?
+                    ).also { newParent ->
+                        newParent.tabs.toMutableList().map {
+                            updateParent(
+                                it,
+                                newParent
+                            ) // all the entries in the parent tabs need to reference their new parent
+                        }
+                    }
+                }
+            },
+            ensureOnHistoryPath = ensureOnHistoryPath
+        )
+    } ?: newItem.populateChildParents()
+
+    return result
+}
+
+private fun <L : Any, T : Any> updateParent(
+    item: Navigation<L, T>,
+    newParent: Navigation<L, T>
+): Navigation<L, T> {
+    when (item) {
+        is BackStack -> item.parent = newParent
+        is EndNode -> item.parent = newParent
+        is TabHost -> item.parent = newParent
+    }
+    return item
+}
+
+internal fun <L : Any, T : Any> Navigation<L, T>.requireParent(): Navigation<L, T> {
+    return parent ?: throw RuntimeException(
+        "We were expecting a non null directParent " +
+                "here. Please file an issue, indicating the function called and the output of " +
+                "toString(diagnostics=true)"
+    )
+}
+
+fun <L : Any, T : Any> BackStack<L, T>.addLocation(location: L): BackStack<L, T> {
+    return copy(
+        stack = stack.toMutableList().also { it.add(endNodeOf(location)) }
+    ).populateChildParents()
+}
+
+fun <L : Any, T : Any> TabHost<L, T>.addLocationToCurrentTab(location: L): TabHost<L, T> {
+    return copy(
+        tabs = tabs.mapIndexed { index, backStack ->
+            if (index == selectedTabHistory.last()) {
+                backStack.addLocation(location)
+            } else {
+                backStack
+            }
+        }
+    ).populateChildParents()
+}
+
+fun <T> List<TabHostLocation<T>>.showSelected(
+    tabHostId: T,
+    index: Int
+): Boolean { //todo review this naming
+    return firstOrNull { it.tabHostId == tabHostId }?.tabIndex == index
+}
+
+suspend fun <L : Any, T : Any> NavigationModel<L, T>.exportState(): String {
+    Fore.d("exportState() ${this.state}")
+    TODO()
+}
+
+suspend fun <L : Any, T : Any> NavigationModel<L, T>.importState(
+    serializedState: String,
+    addToHistory: Boolean = true
+) {
+    Fore.d("importState() addToHistory:$addToHistory")
+    TODO()
+//    val state = serializedState
+//    this.reWriteNavigation(state, addToHistory = addToHistory)
+}
+
+internal fun <L : Any, T : Any> Navigation<L, T>.render(
+    padding: Int = 0,
+    builder: StringBuilder,
+    incDiagnostics: Boolean,
+    current: Boolean = false,
+): StringBuilder {
+    return when (this) {
+        is BackStack -> render(padding, builder, incDiagnostics, current)
+        is EndNode -> render(padding, builder, incDiagnostics, current)
+        is TabHost -> render(padding, builder, incDiagnostics, current)
+    }
+}
+
+private fun <L : Any, T : Any> EndNode<L, T>.render(
+    padding: Int,
+    builder: StringBuilder,
+    incDiagnostics: Boolean,
+    current: Boolean
+): StringBuilder {
+    with(builder) {
+        repeat(padding) { append(" ") }
+        append("endNodeOf(${location!!::class.simpleName})")
+        if (incDiagnostics) {
+            append(" [parent=${parent} child=${child}]")
+            if (current) {
+                append("     <--- Current Item")
+            }
+        } else {
+            if (current) {
+                append(" <---")
+            }
+        }
+        return this
+    }
+}
+
+private fun <L : Any, T : Any> TabHost<L, T>.render(
+    padding: Int,
+    builder: StringBuilder,
+    incDiagnostics: Boolean,
+    current: Boolean
+): StringBuilder {
+    var pad = padding
+    with(builder) {
+        repeat(pad) { append(" ") }
+        pad += padStep
+        append("tabsOf( ")
+        if (incDiagnostics) {
+            append("[tabs=${tabs.size} parent=${parent} child=${child}]")
+        }
+        append("\n")
+        repeat(pad) { append(" ") }
+        append("selectedTabHistory = listOf(")
+        for (index in selectedTabHistory) {
+            append("$index,")
+        }
+        setLength(length - 1)
+        append("),")
+        append("\n")
+        repeat(pad) { append(" ") }
+        append("tabHostId = $tabHostId,\n")
+        if (incDiagnostics) {
+            repeat(pad) { append(" ") }
+            append("backMode = $tabBackModeDefault,\n")
+            repeat(pad) { append(" ") }
+            append("clearToTabRoot = $clearToTabRootDefault,\n")
+        }
+        tabs.forEachIndexed { index, tab ->
+            if (index == selectedTabHistory.last()) {
+                tab.render(pad, builder, incDiagnostics, current)
+            } else {
+                tab.render(pad, builder, incDiagnostics, false)
+            }
+            append(",\n")
+        }
+        setLength(length - 2)
+        append("\n")
+        pad -= padStep
+        repeat(pad) { append(" ") }
+        append(")")
+        return this
+    }
+}
+
+private fun <L : Any, T : Any> BackStack<L, T>.render(
+    padding: Int,
+    builder: StringBuilder,
+    incDiagnostics: Boolean,
+    current: Boolean
+): StringBuilder {
+    var pad = padding
+    with(builder) {
+        repeat(pad) { append(" ") }
+        pad += padStep
+        append("backStackOf( ")
+        if (incDiagnostics) {
+            append("[stackSize=${stack.size} parent=${parent} child=${child}]")
+        }
+        append("\n")
+        val lastIndex = stack.lastIndex
+        stack.forEachIndexed { index, stackItem ->
+            if (index == lastIndex) {
+                stackItem.render(pad, builder, incDiagnostics, current)
+            } else {
+                stackItem.render(pad, builder, incDiagnostics, false)
+            }
+            append(",\n")
+        }
+        setLength(length - 2)
+        append("\n")
+        pad -= padStep
+        repeat(pad) { append(" ") }
+        append(")")
+        return this
+    }
+}
+
+/**
+ * a copy of the navigation item which has had the back operation applied to it, for an
+ * end node, this will be the navigation item with no changes (a back operation is not
+ * valid for this type of navigation item)
+ */
+internal fun <L : Any, T : Any> Navigation<L, T>.createItemNavigatedBackCopy(): Navigation<L, T> {
+    return when (this) {
+        is BackStack -> createItemNavigatedBackCopy()
+        is EndNode -> createItemNavigatedBackCopy()
+        is TabHost -> createItemNavigatedBackCopy()
+    }.populateChildParents()
+}
+
+
+private fun <L : Any, T : Any> EndNode<L, T>.createItemNavigatedBackCopy(): Navigation<L, T> {
+    return this
+}
+
+private fun <L : Any, T : Any> TabHost<L, T>.createItemNavigatedBackCopy(): Navigation<L, T> {
+    return if (specificItemCanNavigateBack()) {
+        copy(selectedTabHistory = selectedTabHistory.toMutableList().also { it.removeLast() })
+    } else this
+}
+
+/**
+ * the returned navigation graph will need to have its parents updated
+ * once the complete nav graph has been recreated by calling populateParents()
+ * on the top level navigation item
+ */
+private fun <L : Any, T : Any> BackStack<L, T>.createItemNavigatedBackCopy(): Navigation<L, T> {
+    return if (specificItemCanNavigateBack()) {
+        copy(stack = stack.toMutableList().also { it.removeLast() })
+    } else this
+}
+
+/**
+ * NOTE: populateParents() MUST be called on the navigation graph BEFORE calling this
+ * function initially (not required for subsequent recursive calls)
+ *
+ * @navigation opportunities for navigating back will be looked for from
+ * this point in the navigation graph, and up via parents (i.e. ignoring children), therefore
+ * clients will typically start by sending currentItem() here
+ *
+ * @returns the complete new navigation graph after the back operation has been
+ * performed or null if it was not possible to navigate further back in the graph
+ */
+fun <L : Any, T : Any> Navigation<L, T>.applyOneStepBackNavigation(): Navigation<L, T>? {
+    Fore.d("calculateBackStep() type:${this::class.simpleName} navigation:${this}")
+    return if (specificItemCanNavigateBack()) {
+        Fore.d("calculateBackStep()... item CAN navigate back")
+        mutateNavigation(
+            oldItem = this,
+            newItem = this.createItemNavigatedBackCopy()
+        )
+    } else { // try to move up the chain
+        Fore.d("calculateBackStep()... item CANNOT navigate back, (need to move up chain to parent) directParent:${parent}")
+        parent?.applyOneStepBackNavigation()
+    }
+}
+
+/**
+ * WARNING: populateParents() MUST be called on the navigation graph AFTER calling
+ * this function as changing parent references is part of the reversal algorithm
+ *
+ * NOTE: populateParents() MUST be called on the navigation graph BEFORE calling this
+ * function initially (not required for subsequent recursive calls)
+ *
+ * @location location to be searched for
+ *
+ * @nav search will be conducted from the currentLocation of this nav graph, and up via
+ * parent relationships, and in the same manner as would a user continually
+ * navigating back from the current item until they exit the app
+ *
+ * @returns a mutated navigation graph containing the searched for location in current position
+ * or null if the location is not found
+ */
+fun <L : Any, T : Any> Navigation<L, T>.reverseToLocation(locationToFind: L): Navigation<L, T>? {
+    Fore.d("reverseToLocation() locationToFind:${locationToFind::class.simpleName} nav:$this")
+    return if (currentLocation()::class.simpleName == locationToFind::class.simpleName) {
+        Fore.d("reverseToLocation()... << MATCHED ${currentLocation()::class.simpleName} >>")
+        this
+    } else {
+        currentItem().applyOneStepBackNavigation()?.reverseToLocation(locationToFind)
+    }
+}
+
+/**
+ * Finds a TabHost if it exists in the navigation graph, even if it is NOT on the back path
+ * of the user. No mutation happens here, a reference to the item is just returned if found, or
+ * null if not found
+ *
+ * NOTE: unlike most functions, this starts at the Home/Exit location and works
+ * downwards through the navigation graph, so this must be initially called on the Home Navigation
+ * item (which will always be a BackStack or a TabHost). If you don't call this function from the
+ * home location, it will throw an exception
+ *
+ * If a BackStack is encountered during a search, each item of that backStack will be checked for
+ * further processing starting at stack[0]
+ *
+ * If a TabHost is encountered during a search (assuming there is no TabHostId match) each tab of
+ * that TabHost is checked in turn for further processing, starting with tab 0
+ *
+ * If there is a TabHost match, that TabHost is returned
+ *
+ * If the search completes with no TabHostId match, null is returned
+ *
+ * @returns a reference to the TabHost as it sits within the navigation graph, or null if the
+ * TabHostId is not present anywhere in the navigation graph
+ */
+fun <L : Any, T : Any> Navigation<L, T>.tabHostFinder(
+    tabHostIdToFind: T,
+    skipParentCheck: Boolean = false
+): TabHost<L, T>? {
+
+    Fore.d("tabHostFinder() tabHostIdToFind:${tabHostIdToFind::class.simpleName} nav:$this")
+
+    if (!skipParentCheck) {
+        require(parent == null) {
+            "This function is intended to be called on the home / top level navigation item, for" +
+                    "which the parent will be null. Try: nav.topParent().tabHostFinder()"
+        }
+    }
+
+    return when (val tabOrBackStack = notEndNode()) {
+        is RestrictedNavigation.NotEndNode.IsBackStack -> {
+            tabOrBackStack.value.stack.firstNotNullOfOrNull {
+                if (it is EndNode) {
+                    Fore.d("tabHostFinder() skip $this")
+                    null
+                } else {
+                    it.tabHostFinder(tabHostIdToFind, true)
+                }
+            }
+        }
+
+        is RestrictedNavigation.NotEndNode.IsTabHost -> {
+            if (tabOrBackStack.value.tabHostId == tabHostIdToFind) {
+                tabOrBackStack.value
+            } else {
+                tabOrBackStack.value.tabs.firstNotNullOfOrNull {
+                    it.tabHostFinder(tabHostIdToFind, true)
+                }
+            }
+        }
+    }
+}
